requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.8.3
firebase-admin==6.4.0
python-dotenv==1.0.0
